implémentation typescript mongoose/ multer et gestion erreurs not found

+++ app.ts
@@ -1,28 +1,21 @@
 import express from 'express';
-import { File } from './models/File';
 import { unlink } from 'fs';
-import { User } from './models/User';
 
-declare global {
-    namespace Express {
-        interface Request {
-            file: File
-        }
-    }
-}
+
 const app = express()
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
 const path = require('path');
-const multer = require('multer')
+import multer, { FileFilterCallback } from 'multer';
+import User from './models/user.model';
 const storage = multer.diskStorage({
-    destination: (req: any, file: any, cb: any) => {
+    destination: (req, file, cb) => {
         cb(null, 'uploads');
     },
-    filename: (req: any, file: any, cb: any) => {
+    filename: (req, file, cb) => {
         cb(null, Date.now() + path.extname(file.originalname));
     }
 });
-const fileFilter = (req: any, file: any, cb: any) => {
+const fileFilter = (req: express.Request, file: Express.Multer.File, cb: FileFilterCallback) => {
     if (file.mimetype == 'image/jpeg' || file.mimetype == 'image/png') {
         cb(null, true);
     } else {
@@ -34,17 +27,6 @@ const upload = multer({ storage: storage, fileFilter: fileFilter })
 
 mongoose.connect('mongodb://localhost:27017/nodejs', { useNewUrlParser: true, useUnifiedTopology: true });
 
-const User = mongoose.model('user', {
-    name: String,
-    surname: String,
-    email: String,
-    password: String,
-    birthDate: Date,
-    civility: String,
-    longitude: Number,
-    latitude: Number,
-    photo: String
-});
 
 app.use(express.json())
 app.use('/static', express.static(__dirname + '/uploads'));
@@ -63,14 +45,15 @@ app.get('/users', async (req, res) => {
 app.get('/users/:id', async (req, res) => {
     try {
         const user = await User.findOne({ _id: req.params.id });
-        res.json(user)
+        user ? res.json(user) : res.status(404).send('Not Found')
+
     } catch (err) {
         res.send(err)
     }
 })
 
-//POST
-app.post('/users', async (req, res) => {
+//POST (PREVIOUS WITHOUT FILE UPLOAD)
+app.post('/old-users', async (req, res) => {
     try {
         const user = await User.create(req.body);
         res.json(user)
@@ -79,11 +62,22 @@ app.post('/users', async (req, res) => {
     }
 })
 
+//POST + UPLOAD
+app.post('/users', upload.single('avatar'), async (req, res) => {
+    try {
+        const user = await User.create({ ...req.body, photo: 'http://localhost:3000/static/' + req.file.filename });
+        res.json(user)
+    } catch (err) {
+        res.status(422)
+        err.errors ? res.send(err) : res.send('A profile pictures is needed')
+    }
+})
+
 //LOGIN
 app.post('/login', async (req, res) => {
     try {
-        const user = await User.findOne({ password: req.body.password, email: req.body.password });
-        res.json(user)
+        const user = await User.findOne({ password: req.body.password, email: req.body.email });
+        user ? res.json(user) : res.status(404).send('Not Found')
     } catch (err) {
         res.send(err)
     }
@@ -109,7 +103,7 @@ app.delete('/users/:id', async (req, res) => {
     }
 })
 
-//UPLOAD IMG 
+//UPLOAD AN IMAGE 
 app.post('/upload', upload.single('avatar'), (req, res) => {
     if (!req.file) {
         console.log("No file received");
@@ -123,7 +117,7 @@ app.post('/upload', upload.single('avatar'), (req, res) => {
     }
 });
 
-//CHANGE IMAGE OF SPECIFIC USER AND DELETE OLD ONE
+//CHANGE IMAGE OF SPECIFIC USER AND DELETE PREVIOUS ONE
 app.post('/upload/:id', upload.single('avatar'), async (req, res) => {
     if (!req.file) {
         console.log("No file received");
@@ -133,13 +127,16 @@ app.post('/upload/:id', upload.single('avatar'), async (req, res) => {
     } else {
         try {
             //DELETE OLD IMAGE
-            const userToDel: User = await User.findOne({ _id: req.params.id })
-            const filename = userToDel.photo.replace('http://localhost:3000/static/', '')
-            unlink('uploads/' + filename, (err) => {
-                if (err) throw err;
-                console.log("deleted file")
-            })
-
+            const userToDel = await User.findOne({ _id: req.params.id })
+            if (userToDel) {
+                const filename = userToDel.photo.replace('http://localhost:3000/static/', '')
+                unlink('uploads/' + filename, (err) => {
+                    if (err) throw err;
+                    console.log("deleted file")
+                })
+            } else {
+                res.json({ status_code: 404, message: "Not Found" })
+            }
             //UPDATE USER WITH NEW PATH
             const user = await User.updateOne({ _id: req.params.id }, { photo: 'http://localhost:3000/static/' + req.file.filename })
             res.json(user)
@@ -149,14 +146,6 @@ app.post('/upload/:id', upload.single('avatar'), async (req, res) => {
     }
 });
 
-//POST / UPLOAD
-app.post('/users/avatar', upload.single('avatar'), async (req, res) => {
-    try {
-        const user = await User.create({ ...req.body, photo: 'http://localhost:3000/static/' + req.file.filename });
-        res.json(user)
-    } catch (err) {
-        res.send(err)
-    }
-})
+
 
 app.listen(3000)
